zucchini: first commit
Zucchini is a simple posix compatible firmware validation framework.

This framework is meant to be use in conjuction of a build-system such
as build but can also be use a standalone.

Tests are run inside a working posix environenment using only shell
script.

The tests are fed to the zucchini script and run in sequence.
Each tests produce a result (pass of fail).

For usage and exmample see the README

Issue: #1582
Change-Id: I5a04910554a12e5af581084d1dc20b5fc5e7eb6b