cukinia: use /proc for _cukinia_process
ps does not have necessary -a and -o options if busybox is not
compiled with the appropriate option. In addition, output of ps
without option is not reliable because the format can differ if run
with the posix or the busybox version.

This commit makes _cukinia_process reliying on data from
/proc/<pid>/status to check if a process name is running and if
optionnally the process owner is matching.

Signed-off-by: Eloi Bail <eloi.bail@savoirfairelinux.com>