cukinia: handle process with arguments
cukinia_process relies on /proc/<pid>/status file to get all running
process names.
In case of process with arguments, the arguments are not written in
status file.
For instance the running process "python3 test.py" will be written
"python3" in status file.

This commit still relies on status file to check the process owner
but use cmdline file which includes the process name concatenated
with all its arguments.

In addition, the use of cmdline file removes the limitation of process
name to 15 chars seen using status file.