refactor cukinia_process around pidof
Several problems around cukinia_process lately. It would return false
positives on more compact busybox configurations, and be extremely
slow due to the scanning of all pids under /proc.

The function is refactoring around pidof, considering it is fairly
standard both in the Busybox and GNU worlds, and it provides a much
more efficient process scanner.