cukinia: ensure that only provided arguments are shifted
"shift" builtin in dash does not support shifting more than the number
of arguments supplied in the argument list. This situation raise a
fatal error that is caught by the shell which eventually interrupts
the script execution.

Only shift up to "$#" arguments to prevent unexpected script endings
when using dash.