cukinia: keep suite name formatting for display
The current version of cukinia strip the suite name given in the conf
file according to this regex : `sed -e 's/[^a-zA-Z0-9_]//g'`.
(keep only alphanumeric character and underscore).

This avoids many problems but also strip interresting character of the
name of the suite itself, particularly spaces. This commit keep an
unformatted version of the test suite and use it only for the field
"name" of the associated junitxml line.