cukinia: add cukinia_kprocess
The current implementation of cukinia_process skips kernel processes.
However, in some cases it can be useful to check if some kernel
processes are running, e.g. watchdogd.
This commit adds the cukinia_kprocess command to check if such a process
is running. No user check is done.
`pgrep -x` is used instead of `pidof` as some versions of `pidof` do
not check kernel processes and do not have the '-w' option do to so.
The differenciation between user and kernel processes is done by
checking if `/proc/pid/cmdline` is empty or not, as is done by `ps`.

SFL: #9476