add on test result statement
This commit adds a new statement `on` that allows to execute a statement
conditionally to the test result.
For the moment, the only conditional statement available is retry.
Retry allows to retry a test a given number of times with an optional
given interval between each attempt.

The `on` and `retry` statement are used as follows:
```
on failure retry <count> after <interval> <command>
```
or
```
on success retry <count> <command>
``

When retry statement is used, an output message is displayed when the
test is retried.
For example for the test:
```
on failure retry 2 after 1s as "Should fail after 2 retries" cukinia_cmd
/bin/false
```
Cukinia will display after the first try:
```
Retry "Should fail after 2 retries" in 1s
```
After the second retry, the classic cukinia fail output will be shown.