cukinia: respect Jenkins Junit xml specification
Respect Jenkins Junit xml specification by adding the following fields
in testsuite:
- id: starts at '0' for the first testsuite and is incremented by 1 for
each following testsuite
- hostname: Host on which the tests were executed.
- system-out: Data that was written to standard output while the test
suite was executed
- system-err: Data that was written to standard error while the test
suite was executed

Also add a cukinia test that will check if the xml output is compliant
with the Junit xml schema. The xsd file used is inspired from:
https://github.com/windyroad/JUnit-Schema/blob/master/JUnit.xsd
but this file has been modified to add the possibility to use
system-out, system-err and properties in the testcase element as Jenkins
Junit xml schema allows it.