cukinia: fix return code with "on retry" statement
When using the "on retry" statement, the return code is always the
one of the first try attempt, even if a later one succeeded.
This is because the _cukinia_retry() function never assigns the return
code of new attempts to the "ret" variable. Thus, even if a test
passed it will return the return code of the first attempt, which might
have failed.
This can be a problem if the rest of a cukinia script uses this return
code.

Fix this behaviour by directly returning the latest attempt return code
when returning from the recursive call stack.