testcases: split failing tests in a separate file
In order to provide CI tests, we will want to simply run PASSing tests
in one go and checking the output is 0. For tests that are expected
to fail however, we'll want to check the output strings.

This split allows to check the success tests with the global return
code, which is no longer poluted by failing tests.