cukinia: add option to generate all output formats
The “all” option can now be passed to the -f option. This allows to
generate csv and xml output without having to run the tests twice.

Since the “-f all” option generates the junitxml file, it also requires
an output file name, given with the -o option. In addition, an extension
(.csv or .xml) will be automatically added to the given file name.