cukinia_contains: add new function to check word lists
This is a common pattern that we used to implement in cukinia_cmd with
egrep. It provides the ability to check if some words are/aren't present
in a whitespace-separated list.

It may be used for instance like cukinia_dmesg, while allowing some
errors/warnings to be ignored based on a filter.