cukinia_matches: add new function to check that all lines match at least one pattern
This is a common pattern we used to define as a cukinia_cmd. It will be
more convenient to have it as a built-in function. Also, this
implementation doesn't rely on egrep, so that's one less dependency.

As illustrated, this is common to make an allowlist of known setuid
binaries or running services.