cukinia: improve message consistency with verbs
Following the previous commit, I noticed that some messages were
not consistent in their use of verbs. Some commands used "running",
others "is running". The syntax was not consistent, and not always
very natural to read.

Aligned all messages to use "is" for state checks, and "does" for
action checks. For example, "Checking file X does exist", or
"Checking process Y is running".